Feature/shipping (#152)
#### What problem is this solving?

When delivery promises are enabled in the account, the merchant may want
to require shoppers to enter address information before adding a product
to the cart.

This PR introduces a new `onClickBehavior` option:
`add-to-cart-and-trigger-shipping-modal`. When set, clicking the
`AddToCartButton` opens a modal prompting for address input if none is
provided. Once the address is entered, the product is added to the cart.

If the address is already set, the product is added without
interruption.

#### How to test it?

<!--- Don't forget to add a link to a Workspace where this branch is
linked -->

[Workspace](https://hiago--vinotecamx.myvtex.com/)

- Try to add a product to the cart by clicking on a `AddToCartButton`.
- Inser any zip code (64633) for example, and select a shipping option.
- The product you clicked will be added to the cart
- Clean your cache if you want to test it again

#### Screenshots or example usage:


https://github.com/user-attachments/assets/d3bd4244-95d3-4992-a897-bb926a5ffa55

---------

Co-authored-by: Victor Moura Cortez <victor.moura@vtex.com>